refactor(domain): consolidate canvas components and refactor structure
- Move canvas-related components to the @/components/organisms/domain/
  namespace for better domain-driven organization.
- Remove deprecated VNodeFloatingEditor molecule component.
- Update structural interfaces and import paths across the codebase
  to align with the new organization.
- Adjust VNodeShield styling to use consistent layout tokens.
