+++ src/components/organisms/domain/canvases/VConceptualEdge.vue
