+++ src/components/organisms/domain/canvases/VEdgeFloatingEditor.vue
