+++ src/components/organisms/domain/canvases/VNodeContainer.vue
