+++ src/components/organisms/domain/canvases/VNodeFloatingEditor.vue
