+++ src/components/organisms/domain/consultation/BlindSpotNavigator.vue
